requirements.txt
llama-index
google-generativeai
typer>=0.9.0
typing_extensions>=4.9.0
pydantic>=2.6.3
python-dotenv>=1.0.1
wikipedia>=1.4.0
streamlit>=1.32.2